Export sso.AuthInfoProcessor
