+++ pkg/auth/dependency/sso/authInfo.go
@@ -1,24 +1,5 @@
 package sso
 
-import (
-	"encoding/json"
-	"io"
-	"strings"
-)
-
-type authInfoProcessor interface {
-	decodeAccessTokenResp(r io.Reader) (AccessTokenResp, error)
-	validateAccessTokenResp(accessTokenResp AccessTokenResp) error
-	processUserID(p map[string]interface{}) string
-	processAuthData(p map[string]interface{}) map[string]interface{}
-}
-
-type defaultAuthInfoProcessor struct{}
-
-func newDefaultAuthInfoProcessor() defaultAuthInfoProcessor {
-	return defaultAuthInfoProcessor{}
-}
-
 type getAuthInfoRequest struct {
 	providerName   string
 	clientID       string
@@ -30,7 +11,7 @@ type getAuthInfoRequest struct {
 	encodedState   string
 	accessTokenURL string
 	userProfileURL string
-	processor      authInfoProcessor
+	processor      AuthInfoProcessor
 }
 
 func (h getAuthInfoRequest) getAuthInfo() (authInfo AuthInfo, err error) {
@@ -46,12 +27,12 @@ func (h getAuthInfoRequest) getAuthInfo() (authInfo AuthInfo, err error) {
 		return
 	}
 
-	accessTokenResp, err := h.processor.decodeAccessTokenResp(r)
+	accessTokenResp, err := h.processor.DecodeAccessTokenResp(r)
 	if err != nil {
 		return
 	}
 
-	err = h.processor.validateAccessTokenResp(accessTokenResp)
+	err = h.processor.ValidateAccessTokenResp(accessTokenResp)
 	if err != nil {
 		return
 	}
@@ -61,9 +42,9 @@ func (h getAuthInfoRequest) getAuthInfo() (authInfo AuthInfo, err error) {
 		return
 	}
 
-	userID := h.processor.processUserID(userProfile)
+	userID := h.processor.ProcessUserID(userProfile)
 	// TODO: process process_userinfo_hook
-	authData := h.processor.processAuthData(userProfile)
+	authData := h.processor.ProcessAuthData(userProfile)
 
 	state, err := DecodeState(h.stateJWTSecret, h.encodedState)
 	if err != nil {
@@ -79,44 +60,5 @@ func (h getAuthInfoRequest) getAuthInfo() (authInfo AuthInfo, err error) {
 		ProviderAuthData:        authData,
 	}
 
-	return
-}
-
-func (d defaultAuthInfoProcessor) decodeAccessTokenResp(r io.Reader) (AccessTokenResp, error) {
-	var accessTokenResp AccessTokenResp
-	err := json.NewDecoder(r).Decode(&accessTokenResp)
-	if err != nil {
-		return accessTokenResp, err
-	}
-	accessTokenResp.Scope = strings.Split(accessTokenResp.RawScope, " ")
-	return accessTokenResp, err
-}
-
-func (d defaultAuthInfoProcessor) validateAccessTokenResp(accessTokenResp AccessTokenResp) error {
-	if accessTokenResp.AccessToken == "" {
-		err := ssoError{
-			code:    MissingAccessToken,
-			message: " Missing access token parameter",
-		}
-		return err
-	}
-
-	return nil
-}
-
-func (d defaultAuthInfoProcessor) processUserID(userProfile map[string]interface{}) string {
-	id, ok := userProfile["id"].(string)
-	if !ok {
-		return ""
-	}
-	return id
-}
-
-func (d defaultAuthInfoProcessor) processAuthData(userProfile map[string]interface{}) (authData map[string]interface{}) {
-	authData = make(map[string]interface{})
-	email, ok := userProfile["email"].(string)
-	if ok {
-		authData["email"] = email
-	}
 	return
 }
